requirements.txt
# Servidor e Framework Core
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Validação de Dados e Ambiente
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Utilitários
httpx==0.27.0